Rename search bar to controls
